[#4417] SFLQ Create/Pick added to LeaseMgrs
Added new SQL statements and functions to
create and pick from SFLQ pools to
MySQL and PostgreSQL lease managers

/src/hooks/dhcp/pgsql/pgsql_lease_mgr.cc
    New SQL statements:
    SFLQ_CREATE_POOL4
    SFLQ_PICK_FREE_LEASE4
    SFLQ_CREATE_POOL6
    SFLQ_PICK_FREE_LEASE6

    New functions:
    PgSqlLeaseMgr::sflqCreateFlqPool4()
    PgSqlLeaseMgr::sflqPickFreeLease4()
    PgSqlLeaseMgr::sflqCreateFlqPool6()
    PgSqlLeaseMgr::sflqPickFreeLease6()

/src/hooks/dhcp/pgsql/tests/pgsql_lease_mgr_unittest.cc
    New tests:
    TEST_F(PgSqlLeaseMgrTest, testSFLQ4)
    TEST_F(PgSqlLeaseMgrTest, testSFLQ6)

/src/hooks/dhcp/mysql/mysql_lease_mgr.*
    New SQL statements:
    MySqlLeaseMgr::SFLQ_CREATE_FLQ_POOL4
    MySqlLeaseMgr::SFLQ_PICK_FREE_LEASE4
    MySqlLeaseMgr::SFLQ_CREATE_FLQ_POOL6
    MySqlLeaseMgr::SFLQ_PICK_FREE_LEASE6

    New functions:
    MySqlLeaseMgr::sflqCreateFlqPool4()
    MySqlLeaseMgr::sflqPickFreeLease4()
    MySqlLeaseMgr::sflqCreateFlqPool6()
    MySqlLeaseMgr::sflqPickFreeLease6()

/src/hooks/dhcp/mysql/tests/mysql_lease_mgr_unittest.cc
    New tests:
    TEST_F(MySqlLeaseMgrTest, testSFLQ4)
    TEST_F(MySqlLeaseMgrTest, testSFLQ6)

/src/lib/dhcpsrv/lease_mgr.*
    New Functions:
    LeaseMgr::sflqCreateFlqPool4()
    LeaseMgr::sflqPickFreeLease4()
    LeaseMgr::sflqCreateFlqPool6()
    LeaseMgr::sflqPickFreeLease6()

/src/lib/dhcpsrv/testutils/generic_lease_mgr_unittest.*
    New tests:
    GenericLeaseMgrTest::testSFLQ4()
    GenericLeaseMgrTest::testSFLQ6()

+++ src/hooks/dhcp/mysql/tests/mysql_lease_mgr_unittest.cc
@@ -81,8 +81,13 @@ public:
     /// @brief Destroys the LM and the schema.
     void destroyTest() {
         LeaseMgrFactory::destroy();
-        // If data wipe enabled, delete transient data otherwise destroy the schema
-        destroyMySQLSchema();
+        if (getenv("KEA_UNIT_TEST_KEEP_SCHEMA")) {
+            // Leaves schema intact for post-test debugging.
+            std::cout << "KEA_UNIT_TEST_KEEP_SCHEMA set" << std::endl;
+        } else {
+            // If data wipe enabled, delete transient data otherwise destroy the schema
+            destroyMySQLSchema();
+        }
 
         // Disable Multi-Threading.
         MultiThreadingMgr::instance().setMode(false);
@@ -1435,6 +1440,14 @@ TEST_F(MySqlLeaseMgrTest, updateStatsOn6DifferentSubnet) {
 
 TEST_F(MySqlLeaseMgrTest, updateStatsOn6DifferentSubnetPD) { 
     testUpdateStatsOn6DifferentSubnetPD();
+}
+
+TEST_F(MySqlLeaseMgrTest, testSFLQ4) { 
+    testSFLQ4();
+}
+
+TEST_F(MySqlLeaseMgrTest, testSFLQ6) { 
+    testSFLQ6();
 }
 
 /// @brief Test fixture class for testing @ref CfgDbAccessTest using MySQL
